+++ src/ui/log_tab.rs
@@ -209,6 +209,7 @@ impl<'a> LogTab<'a> {
         Ok(Some(AppAction::SetPopup(Box::new(log_context_menu(
             get_env().jj_config.clone(),
             anchor,
+            &self.selection(),
             &self.head,
             &self.marked(),
         )?))))